ログインフラグ実装
